+++ crates/pray-cli/src/command.rs
@@ -94,4 +94,7 @@ pub(crate) enum Command {
     },
     Upgrade,
     Version,
+    Completion {
+        shell: String,
+    },
 }